# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    ""
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/addFavouriteProduct": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "添加商品到我的最愛",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "productId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/getFavouriteProduct": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "添加商品到我的最愛",
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/hello": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "驗證功能測試路由",
                "responses": {
                    "200": {
                        "description": "{\"code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "用戶登出",
                "responses": {
                    "200": {
                        "description": "{\"code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "用戶登入",
                "parameters": [
                    {
                        "description": "使用者名稱",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/createProduct": {
            "post": {
                "tags": [
                    "商品"
                ],
                "summary": "新增商品",
                "parameters": [
                    {
                        "description": " ",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateProductType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " ",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/findProduct": {
            "get": {
                "tags": [
                    "商品"
                ],
                "summary": "查詢產品資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "productId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": " ",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/getProduct": {
            "get": {
                "tags": [
                    "商品"
                ],
                "summary": "取得產品資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": " ",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/updateProduct": {
            "post": {
                "tags": [
                    "商品"
                ],
                "summary": "更新產品資訊",
                "parameters": [
                    {
                        "description": " ",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " ",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/createUser": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "新增用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者名稱",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "使用者密碼",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "使用者確認密碼",
                        "name": "repassword",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/deleteUser": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "刪除用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者ID",
                        "name": "id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/updateUser": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "編輯用戶",
                "parameters": [
                    {
                        "description": " ",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/uploadUserAvatar": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "用戶照片上傳",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "使用這ID",
                        "name": "userId",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "1:1圖片",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateProductType": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "productImg": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variation": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateProductVariationType"
                    }
                }
            }
        },
        "models.CreateProductVariationType": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                },
                "specifications": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "models.ProductInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "productImg": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variation": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateProductVariationType"
                    }
                }
            }
        },
        "models.UserInfo": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sex": {
                    "description": "Birthday  *time.Time",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}